+++ tests/Views/Filters/NumberFilterTest.php
@@ -158,13 +158,16 @@ final class NumberFilterTest extends TestCase
         $filter = NumberFilter::make('BreedID');
         $this->assertSame(123, $filter->validate(123));
         $this->assertSame(123.51, $filter->validate(123.51));
-        $this->assertFalse($filter->validate('123'));
+        $this->assertSame(123, $filter->validate('123'));
+        $this->assertSame(123.51, $filter->validate('123.51'));
+
     }
 
     public function test_can_get_if_number_filter_empty(): void
     {
         $filter = NumberFilter::make('Active');
         $this->assertTrue($filter->isEmpty(''));
+        $this->assertTrue($filter->isEmpty('q'));
         $this->assertFalse($filter->isEmpty('123'));
     }